+++ core/data-layer/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-sqlx = { version = "0.7.1", features = [
+sqlx = { version = "0.6.3", features = [
   "postgres",
   "mysql",
   "sqlite",
@@ -15,7 +15,6 @@ sqlx = { version = "0.7.1", features = [
 ] }
 chrono = { version = "0.4.26", default-features = false, features = [
   "clock",
-  "std",
   "serde",
 ] }
 anyhow = "1.0.70"
