+++ core/wave-autoscale/Cargo.toml
@@ -29,10 +29,7 @@ aws-sdk-autoscaling = "0.25.1"
 aws-sdk-lambda = "0.25.1"
 aws-sdk-cloudwatch = "0.25.1"
 quick-js = { version = "0.4.1" }
-chrono = { version = "0.4.26", default-features = false, features = [
-  "clock",
-  "std",
-] }
+chrono = { version = "0.4.26", default-features = false, features = ["clock"] }
 aws-smithy-types-convert = { version = "0.55.0", features = ["convert-chrono"] }
 aws-smithy-types = "0.55.0"
 uuid = { version = "1.3.1", features = ["v4", "fast-rng", "macro-diagnostics"] }
