+++ core/wave-cli/tests/integration_test.rs
@@ -1,0 +1,6 @@
+mod integration_test {
+    #[test]
+    fn test_() {
+        // TODO: Implement integration test
+    }
+}
